+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+* Add tests for common.util.time
 * Display total hours played on homepage
 * Add format_duration to common.util.time
 * Allow deleting sessions